rag_agent/requirements.txt
langchain
langchain-google-genai
langchain-core
langchain-community
langgraph
python-dotenv
requests
faiss-cpu
pypdf
 